refactor: file change; trial_parameters -> parameters.py
